# http: optimize header map implementation for improved performance