@ggreenway 2025-04-28
I'd like to see a performance comparison that mimics real-world access patterns. Specifically, compare lots of use of core headers like `:path` and `:authority` and the other headers that envoy must access for routing/processing. In the current implementation there are two classes of headers, and all of the core-protocol headers are in the "fast" path. So I'd like to see comparison of that.

Also, I'd like to see results in a full Envoy end-to-end test to see how much of a difference this change makes in overall performance.